requirements.txt
numpy==2.3.1
Pillow==11.2.1
streamlit==1.46.0
torch==2.7.1
